requirements.txt
certifi==2020.6.20
cycler==0.10.0
dask==2.30.0
kiwisolver==1.3.0
lz4==3.1.0
matplotlib==3.3.2
natsort==7.0.1
numpy==1.19.3
opencv-python==4.4.0.44
pandas==1.1.3
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
tqdm==4.51.0
wincertstore==0.2
